pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "banyan-extract"
version = "0.8.0"
license = { text = "BSD 3" }
authors = [
    { name="Sandia National Laboratories" },
]
requires-python = ">= 3.10"

dependencies = [
    "python-pptx",
    "pandas",
    "pdf2image",
    "python-dotenv"
    ]

[project.optional-dependencies]
marker = ["marker-pdf", "surya-ocr"]
    
nemotronparse = ["openai"]

[project.scripts]
banyan-extract = "banyan_extract.cli:main"
